backend/requirements.txt
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.0.0
drf-spectacular>=0.26.0
mysqlclient>=2.2.0
Pillow>=10.0.0
gunicorn>=21.2.0
whitenoise>=6.6.0
dj-database-url>=2.1.0
psycopg2-binary>=2.9.9

Faker>=20.0.0

# Cloudflare R2 (S3-compatible storage)
boto3>=1.34.0
django-storages>=1.14.0
